Migrate game class to correct file location, as well as game spec
